login-history-feature-added

+++ AuthService/src/main/java/com/healthcare/repository/ILoginHistoryRepository.java
@@ -4,5 +4,5 @@ package com.healthcare.repository;
 import com.healthcare.model.entity.LoginHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ILoginHistoryRepository extends JpaRepository<LoginHistory,Integer> {
+public interface ILoginHistoryRepository extends JpaRepository<LoginHistory,Long> {
 }
